internal/testutils/netnsutils: move utils that were not used on Windows
