+++ internal/testutils/netnsutils/context_windows.go
@@ -1,19 +1,8 @@
 package netnsutils
 
-import (
-	"testing"
+import "testing"
 
-	"github.com/docker/docker/internal/testutils"
-)
-
-type OSContext struct{}
-
-func SetupTestOSContextEx(*testing.T) *OSContext {
-	return nil
-}
-
-func (*OSContext) Cleanup(t *testing.T) {}
-
-func (*OSContext) Set() (func(testutils.Logger), error) {
-	return func(testutils.Logger) {}, nil
+// SetupTestOSContext is a no-op on Windows.
+func SetupTestOSContext(*testing.T) func() {
+	return func() {}
 }
